+++ apps/simulator/src/components/Header/Header.tsx
@@ -14,6 +14,7 @@ import { usePreloadedGlobals } from "@/core/context/PreloadedGlobalsContext";
 import { useCountdown } from '../../../../../packages/utils/hooks/useCountdown'
 import LoadingBar from 'react-top-loading-bar'
 import { useEffect, useState } from 'react';
+import { useProgress } from '@/core/context/ProgressContext';
 
 
 
@@ -29,22 +30,11 @@ const buttonStyle = {
 
 export const Header: React.FC = () => {
   const { header } = usePreloadedGlobals()
-  const [progress, setProgress] = useState<number>(10)
+  const { progress } = useProgress()
+
   const headerTimeRemaining = header[0]?.timeRemaining ?? null;
   const duration = header[0]?.duration ?? null;
   const { timeRemaining, duration: timeDuration } = useCountdown({ timeRemaining: "04:00:00", duration: "01:00:00" })
-
-  useEffect(() => {
-    const intervalId = setInterval(() => {
-      if (progress <= 100) {
-        setProgress(prev => prev + prev)
-        return
-      }
-      clearInterval(intervalId)
-    }, 400);
-    return () => clearInterval(intervalId)
-  }, [progress])
-
 
   return (
     <Box sx={{ flexGrow: 1 }}>
